Prefer orderless completion style over others
The reason is that while filtering Org headlines, sometimes one of the
other completion styles matches, hiding other possible matches that
would have been found by the `orderless` style.  This can happen when
spaces are included in Org headlines (which is usually the case) and
those spaces match the search pattern.  In this case, the position of
point is also relevant, which is not something I want.

It would be better to know which completion category is use in this case
and override the completion style for this only, but so far I haven't
found it.  Continue looking …