Use basic completion before orderless
Using orderless by default get's in the way of “usual” code completion
and shell completion, where prefix completion is the expected
behavior (for me).  So let's try to use basic completion first, and try
to rely on orderless completion to kick in as soon as spaces are
involved (which should not happen for shell and code completion).

The completion colors should also help to see whether we are still using
basic completion or orderless completion.  Let's try this for now.