Do not enable proced auto-updating on Windows
This seems to be slow, causing proced to take more time updating a
buffer than waiting the one second until the next update – resulting in
an unresponsive proced buffer.

Note: I am not sure whether this setting is correct, as `(not
on-windows)` will not be evaluated when `use-package` calls
`custom-theme-set-variables`.  However, I think it will be evaluated
eventually, and that should be all that's needed.