Remove explicit final newlines in Org capture templates
`org-capture-fill-template` unconditionally adds a final newline to each
template.  This caused extra empty lines with the old templates when
nothing is inserted at point (`%?`), because in this case empty lines
before and after point were present, with nothing in between.

Those extra empty lines are gone now, and point is positioned after the
final non-blank character.  When extra empty lines are needed, they have
to be inserted manually.