Let Git ignore persistent variables
The values persisted here seem to be irrelevant across Emacs
installations, so there is no need to track those.  It also does not
seem to officially customize the location that persist.el uses for
persisting variables: the variable that stores the directory path is
called `persist--directory-location`, indicating a private variable.