Fix handling of last day in workload overview report
Previously, the last day added had been removed on the assumption that
it was the frist day after `end-date` (after which the while loop
aborts).  However, when this last day is filtered because of
`:skip-matches`, this would result in accidentally removing a valid day
from the overview time period.  This has been fixed by checking for this
very case, and keeping the last day added in case it is still within the
requested time period.