Resolve link generation conflict in bookmarked files
When a file has a bookmark, generating a link in that file (e.g. to the
line at point) results in a query from Org which link generating
function to use.  This is annoying and interferes with the usual flow.
To remedy this, we simply try not to be too smart and just generate
links to bookmarks when in the bookmark menu.