Avoid customization conflicts between Customize and use-package
To this end, do not use :custom in use-package declaration, because
according to the documentation[1], `:custom` declarations can conflict
with settings done via `customize-option`.  However, values set directly
in the configuration files should always be allowed to be overwritten by
custom settings done by the user – at least that's the general approach
of this Emacs configuration.

[1]: https://www.gnu.org/software/emacs/manual/html_mono/use-package.html#User-options